Add ExternFuncArgument.h
Most that need Function don't need ExternFuncArgument,  but putting both in Function.h adds extra includes (eg Buffer.h) in places that it isn't needed. Split it out into a separate file and forward-declared in all headers (except Param.h, which will be addressed in the future).
