+++ src/FindCalls.cpp
@@ -1,5 +1,6 @@
 #include "FindCalls.h"
 
+#include "ExternFuncArgument.h"
 #include "Function.h"
 #include "IRVisitor.h"
 #include <utility>
