+++ src/Func.h
@@ -2356,9 +2356,7 @@ public:
     // @}
 
     /** Use a Func as an argument to an external stage. */
-    operator ExternFuncArgument() const {
-        return ExternFuncArgument(func);
-    }
+    operator ExternFuncArgument() const;
 
     /** Infer the arguments to the Func, sorted into a canonical order:
      * all buffers (sorted alphabetically by name), followed by all non-buffers
