+++ src/Inline.cpp
@@ -2,6 +2,7 @@
 
 #include "CSE.h"
 #include "Debug.h"
+#include "ExternFuncArgument.h"
 #include "IRMutator.h"
 #include "IROperator.h"
 #include "IRPrinter.h"
